+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Playwright tests to sort search results (DR-3659)
 - Added Image ID to Items page (DR-3695)
 - Added Playwright test feedback button to the home page (DR-3658)
+- Added Playwright  test where a user clicks Divisions from home page navigation menu (DR-3655)
+- Added Playwright updated test to verify navigation menu on Divisons page (DR-3655)
+- Added Playwright test to verify Division's name headings and the see more links for each division (DR-3655)
 
 ### Fixed
 - Fixed Order Print button (DR-3486)
@@ -21,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated breadcrumb to include Division and Collection information (DR-3701)
 - Updated Collections page to bolden the Polonsky note (DR-3705)
 - Updated `fetch` calls to include original client IP (DR-3754)
+- Updated Item page to use React Player for AV materials
 
 ## [0.4.3] 2025-06-17
 ### Added
@@ -32,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Playwright test to confirm visibility of search result filters (DR-3659)
 - Added Playwright test to filter search results with dropdowns (DR-3659)
 - Added GitHub Actions workflow to run Playwright tests on PRs (DR-3721)
+- Added Playwright tests to clear search results filters (DR-3659)
+
 
 ## [0.4.2] 2025-05-22